Refactor user module
